Fix MRect
Float -> CGFloat

+++ Mallet/UI/MEditorFrameView.swift
@@ -8,7 +8,7 @@
 
 import SwiftUI
 
-fileprivate enum FactorValue: Float {
+fileprivate enum FactorValue: CGFloat {
     case positive = 1
     case zero = 0
     case negative = -1
@@ -29,9 +29,9 @@ fileprivate struct MEditorFrameDot: View {
 
     private let dotColor = Color.blue
 
-    private let minWidth: Float = 10
+    private let minWidth: CGFloat = 10
 
-    private let minHeight: Float = 10
+    private let minHeight: CGFloat = 10
 
     var body: some View {
         Circle()
@@ -42,12 +42,12 @@ fileprivate struct MEditorFrameDot: View {
                     .foregroundColor(Color.white.opacity(0.001))
                     .frame(width: 25, height: 25)
                     .gesture(DragGesture()
-                            .onChanged { value in
-                                self.frame.x += min(self.frameFactors.x.rawValue * Float(value.translation.width), self.frame.width - self.minWidth)
-                                self.frame.y += min(self.frameFactors.y.rawValue * Float(value.translation.height), self.frame.height - self.minHeight)
-                                self.frame.width = max(self.frame.width + self.frameFactors.width.rawValue * Float(value.translation.width), self.minWidth)
-                                self.frame.height = max(self.frame.height + self.frameFactors.height.rawValue * Float(value.translation.height), self.minHeight)
-                        }
+                                 .onChanged { value in
+                                     self.frame.x += min(self.frameFactors.x.rawValue * value.translation.width, self.frame.width - self.minWidth)
+                                     self.frame.y += min(self.frameFactors.y.rawValue * value.translation.height, self.frame.height - self.minHeight)
+                                     self.frame.width = max(self.frame.width + self.frameFactors.width.rawValue * value.translation.width, self.minWidth)
+                                     self.frame.height = max(self.frame.height + self.frameFactors.height.rawValue * value.translation.height, self.minHeight)
+                                 }
                     )
             )
     }
@@ -79,52 +79,52 @@ struct MEditorFrameView<Content>: View where Content: View {
 
     var body: some View {
         content()
-            .frame(width: CGFloat(frame.width), height: CGFloat(frame.height))
+            .frame(width: frame.width, height: frame.height)
             .background(Color.black.opacity(0.03))
-            .position(x: CGFloat(frame.midX), y: CGFloat(frame.midY))
+            .position(x: frame.midX, y: frame.midY)
             .overlay(
                 Group {
                     ZStack {
                         Rectangle()
                             .foregroundColor(Color.white.opacity(0.001))
                             .gesture(DragGesture()
-                                    .onChanged { value in
-                                        self.frame.x += Float(value.translation.width)
-                                        self.frame.y += Float(value.translation.height)
-                                        self.setSelectedUIID()
-                                })
-                            .frame(width: CGFloat(self.frame.width) + self.borderWidth, height: CGFloat(self.frame.height) + self.borderWidth)
-                            .position(x: CGFloat(self.frame.midX), y: CGFloat(self.frame.midY))
+                                         .onChanged { value in
+                                             self.frame.x += value.translation.width
+                                             self.frame.y += value.translation.height
+                                             self.setSelectedUIID()
+                                         })
+                            .frame(width: self.frame.width + self.borderWidth, height: self.frame.height + self.borderWidth)
+                            .position(x: self.frame.midX, y: self.frame.midY)
 
                         Group {
                             if self.uiData.uiID == self.selectedUIID {
                                 Rectangle()
                                     .foregroundColor(.clear)
                                     .border(Color.gray, width: self.borderWidth)
-                                    .frame(width: CGFloat(self.frame.width) + self.borderWidth, height: CGFloat(self.frame.height) + self.borderWidth)
-                                    .position(x: CGFloat(self.frame.midX), y: CGFloat(self.frame.midY))
+                                    .frame(width: self.frame.width + self.borderWidth, height: self.frame.height + self.borderWidth)
+                                    .position(x: self.frame.midX, y: self.frame.midY)
 
 
                                 Group {
                                     MEditorFrameDot(frame: self.$frame, frameFactors: FrameFactors(x: .positive, y: .positive, width: .negative, height: .negative))
-                                        .position(x: CGFloat(self.frame.x), y: CGFloat(self.frame.y))
+                                        .position(x: self.frame.x, y: self.frame.y)
                                     MEditorFrameDot(frame: self.$frame, frameFactors: FrameFactors(x: .zero, y: .positive, width: .positive, height: .negative))
-                                        .position(x: CGFloat(self.frame.x) + CGFloat(self.frame.width), y: CGFloat(self.frame.y))
+                                        .position(x: self.frame.x + self.frame.width, y: self.frame.y)
                                     MEditorFrameDot(frame: self.$frame, frameFactors: FrameFactors(x: .zero, y: .zero, width: .positive, height: .positive))
-                                        .position(x: CGFloat(self.frame.x) + CGFloat(self.frame.width), y: CGFloat(self.frame.y) + CGFloat(self.frame.height))
+                                        .position(x: self.frame.x + self.frame.width, y: self.frame.y + self.frame.height)
                                     MEditorFrameDot(frame: self.$frame, frameFactors: FrameFactors(x: .positive, y: .zero, width: .negative, height: .positive))
-                                        .position(x: CGFloat(self.frame.x), y: CGFloat(self.frame.y) + CGFloat(self.frame.height))
+                                        .position(x: self.frame.x, y: self.frame.y + self.frame.height)
                                 }
 
                                 Group {
                                     MEditorFrameDot(frame: self.$frame, frameFactors: FrameFactors(x: .zero, y: .positive, width: .zero, height: .negative))
-                                        .position(x: CGFloat(self.frame.x) + CGFloat(self.frame.width) / 2, y: CGFloat(self.frame.y))
+                                        .position(x: self.frame.x + self.frame.width / 2, y: self.frame.y)
                                     MEditorFrameDot(frame: self.$frame, frameFactors: FrameFactors(x: .zero, y: .zero, width: .positive, height: .zero))
-                                        .position(x: CGFloat(self.frame.x) + CGFloat(self.frame.width), y: CGFloat(self.frame.y) + CGFloat(self.frame.height) / 2)
+                                        .position(x: self.frame.x + self.frame.width, y: self.frame.y + self.frame.height / 2)
                                     MEditorFrameDot(frame: self.$frame, frameFactors: FrameFactors(x: .zero, y: .zero, width: .zero, height: .positive))
-                                        .position(x: CGFloat(self.frame.x) + CGFloat(self.frame.width) / 2, y: CGFloat(self.frame.y) + CGFloat(self.frame.height))
+                                        .position(x: self.frame.x + self.frame.width / 2, y: self.frame.y + self.frame.height)
                                     MEditorFrameDot(frame: self.$frame, frameFactors: FrameFactors(x: .positive, y: .zero, width: .negative, height: .zero))
-                                        .position(x: CGFloat(self.frame.x), y: CGFloat(self.frame.y) + CGFloat(self.frame.height) / 2)
+                                        .position(x: self.frame.x, y: self.frame.y + self.frame.height / 2)
                                 }
                             }
                         }
@@ -132,9 +132,9 @@ struct MEditorFrameView<Content>: View where Content: View {
                 }
 
                     .gesture(TapGesture()
-                            .onEnded {
-                                self.setSelectedUIID()
-                        }
+                                 .onEnded {
+                                     self.setSelectedUIID()
+                                 }
                     )
             )
     }
